Added and Updated scripts.js page with new loadList() and loadDetails() function
If applied, this commit will have added and update the scripts.js page with a loadList() function as a return key that uses fetch() to GET  the complete list of pokemon from: "https://pokeapi.co/api/v2/pokemon/?limit=150".
A loadDetails() function that should GET the Pokémon details using the URL from the Pokémon object in the parameter.
The showDetails() function was edited to load the Pokemon details from the API calling the loadDetails() function