Updated scripts.js file with a new showModal() function, a new hideModal() function and updated showDetails() function.
If applied, this commit will have added a new showModal() function that generates the information of each pokemon when 'click' the 'button' on .pokemon-list.
Was created a new hideModal() to close the modal when 'click' in the 'X' button or outside of the modal and press 'Escape' Keyword.
The showDetails() function was updated adding the showModal() function in the loadDetails() function to generate the information in the modal.