Updated styles.css page with new #modal-container, #modal-container.is-visible, .modal, .modal-close, .modal h1, .modal p and modal img. Changes in .pokemon-list, button-style and @media
If applied, this commit will have #modal-container, #modal-container.is-visible, .modal, .modal-close, .modal h1, .model p, .modal img that are focused in the new modal created in JavaScript (showModal function). All this classes and id create the first basic styling in the modal.
The .pokemon-list had some adjustments with the number of columns, color and align.
The button-style was resized.
All the design was updated in @media