Added new access with Bootstrap Framework adding a new Navigation menu and a new Modal container. Added new button in footer and changed the main title and pictures
If applied, this commit will have added a new access with Bootstrap adding a Navigation menu that contains a new icon(image) in the left side that have the access to index.html after clicked. This is followed by the official web site link, another link for curiosities and in the right side a form search/filter for the pokemon.
New Modal from Bootstrap added.