Changed showModal function in order to be adapted to the new Modal from Bootstrap. Added new search filter for the new navigation menu from Bootstrap
If applied, this commit will have changed the showModal function to interact with the new Modal from Bootstrap. For this was use JQuery language that intercept the modal classes ('.modal-body' and '.modal-title') and give them the new information to be displayed, in the case the pokemon name, height, weight, image and type.
As well was changed the pokemonRepository.loadList() function to load the the list as well in the input search/filter from the navigation menu. For this was selected the .searchForm(that is connected to he form)and created an eventListener on it that selectes the #searchInput(that is connected to the input value) and create a loop with conditions that allow(or not) a search result.