Removed hideModal() function and two eventListeners that allowed to close the old pressing 'Escape' Keyboard and 'Clicking' in the 'modalContainer'
If applied, this commit will have removed the function hideModal() and two EventListeners that use to allow to close the old Modal (Modal generated with showModal() function that was substituted with Bootstrap Modal).
The first eventListener was able to close the Modal through 'clicking' in the 'modalContainer'(the area outside the modal).
The second one through pressing 'Escape' keyboard.
All of them got removed because, the new Bootstrap Modal bring together this code (similar, with same propose) already done.

Old comments where deleted as well to not confuse new information.