Changed css and scripts file source for the minifier file.
If applied, this commit will have changed the css and scripts source in index.html to a minifier file that allows the files become lighter. The minifier files are just used in the when the project is ready. For further changes we use the previous 'styles.css' and 'scripts.js' files again.